Extend grub-efi to also cover arm64/ia64/arm
Tools like simple-cdd ensures the presence of grub-efi on the generated
ISO image but don't have a simple way to to deal with architecture
specific packages.

Thus I'd like to suggest to keep grub-efi not as a transitional package
but as normal dummy package depending on the package most likely to be
appropriate for each architecture.

Closes: #981819